src/test/java/com/poc/microservice/services/DummyServiceTest.java
package com.poc.microservice.services;

import com.poc.microservice.clients.DummyClient;
import com.poc.microservice.entities.Dummy;
import com.poc.microservice.repositories.DummyRepository;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.Mock;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.when;
import static org.mockito.MockitoAnnotations.initMocks;

class DummyServiceTest {

    @Mock
    private DummyClient mockDummyClient;
    @Mock
    private DummyRepository mockDummyRepository;

    private DummyService dummyServiceUnderTest;

    @BeforeEach
    void setUp() {
        initMocks(this);
        dummyServiceUnderTest = new DummyService(mockDummyClient, mockDummyRepository);
    }

    @Test
    void testGetDummyById() {
        // Setup

        // Configure DummyRepository.getOne(...).
        final Dummy dummy = new Dummy();
        dummy.setId(0L);
        dummy.setDummy("dummy");
        when(mockDummyRepository.getOne(0L)).thenReturn(dummy);

        // Run the test
        final String result = dummyServiceUnderTest.getDummyById(0L);

        // Verify the results
        assertEquals("dummy", result);
    }

    @Test
    void testGetDummyByIdFromClient() {
        // Setup
        when(mockDummyClient.getById(0L)).thenReturn("result");

        // Run the test
        final String result = dummyServiceUnderTest.getDummyByIdFromClient(0L);

        // Verify the results
        assertEquals("result", result);
    }
}
